[Server] Move Topic to separate file
